Move commons-compress to dependency management
this dependency had conflicts between xml-parser (1.12) and commons-rest (1.19), and the xml-parser version was the one being used. By putting the dependency in the dependencyManagement section of the pom, it forces this specific version to be used. The reason 1.18 is used, is because this is the version used by the excel dependencies.