Blocker issue: https://www.ebi.ac.uk/panda/jira/browse/TRM-27473
Streaming requests were causing REST application to become unresponsive. This commit addresses this problem. We will however, need to tidy the logic for this in future. Main goal of this task though is to fix the issue.

Stress test showed 500 users for 6 hours, caused no problem in the application. Many 429s, but that's fine. The application did not die and was responsive after the test finished.